Refactor- Media Details
